Reduce grammar to a minimum
Our current focus should on https://github.com/tarides/client-semgrep/issues/85

> We need a tree-sitter-gosu repo that includes a minimal tree-sitter
grammar, capable of parsing the simplest valid .gs file.

It is not useful to our aims if we can parse all possible .gs files,
with all of its various constructs, if we cannot get fully end-to-end
parsing into Semgrep. The more complex we make the CST, the more
difficult it will be make the initial integration.

For this reason, we must resist the completionist urge, and focus on the
minimal working thing to get the integration set up.