Improve representation of fields
This is a great example of how/why we ought not to transcribe the EBNF
grammar. When we map to the AST, we need to place the modifier together
with the feildDefn, but if we follow the EBNF it is under the
declaration node. When converting to the AST, this would require us to
construct the fieldDefn, then grab the modifier from the parent node and
insert it into the fieldDefn after the fact. But if we move the modifier
into the fieldDefn, we have it when we need it. Ditto with the
semicolon.

This is also how the grammar is approached in the Java tree-sitter
grammar:
https://github.com/tree-sitter/tree-sitter-java/blob/12c4848bc6cf2660d80b435f8d8de1a9d4676d04/grammar.js#L1015-L1020